+++ examples/electric-sql/.electric_migrations_tmp_mwrE2v/migrations/20231018174208_471/migration.sql
@@ -1,5 +1,0 @@
-CREATE TABLE "users" (
-  "id" TEXT NOT NULL,
-  "name" TEXT NOT NULL,
-  CONSTRAINT "users_pkey" PRIMARY KEY ("id")
-) WITHOUT ROWID;